テストケース修正
