+++ tests/Feature/Routes/HomeRoutesTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Tests\Unit\Routes;
+namespace Tests\Feature\Routes;
 
-use Tests\TestCase;
 use App\User;
+use Tests\TestCase;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
-class TaskRoutesTest extends TestCase
+class HomeRoutesTest extends TestCase
 {
 
     use DatabaseTransactions;
@@ -35,11 +35,11 @@ class TaskRoutesTest extends TestCase
      *
      * @test
      */
-    public function it_can_access_the_task_page()
+    public function it_can_access_the_home_page()
     {
-        $response = $this->actingAs($this->user)->get('/task');
+        $response = $this->actingAs($this->user)->get('/home');
         $response->assertStatus(200);
-        $response->assertViewHasAll(['tasks', 'priorities']);
+        $response->assertViewHasAll(['myProjects', 'userProjects', 'userProjectCount', 'userProgressAreaspline']);
     }
 
     /**
@@ -47,11 +47,11 @@ class TaskRoutesTest extends TestCase
      *
      * @test
      */
-    public function it_can_access_the_unfinished_task_page()
+    public function it_can_access_the_home_project_page()
     {
-        $response = $this->actingAs($this->user)->get('/task/unfinished');
+        $response = $this->actingAs($this->user)->get('/home/project');
         $response->assertStatus(200);
-        $response->assertViewHasAll(['tasks', 'priorities']);
+        $response->assertViewHasAll(['myProjects', 'userProjects', 'userProjectCount', 'userProgressAreaspline']);
     }
 
     /**
@@ -59,10 +59,10 @@ class TaskRoutesTest extends TestCase
      *
      * @test
      */
-    public function it_can_access_the_finished_task_page()
+    public function it_can_access_the_user_profile_page()
     {
-        $response = $this->actingAs($this->user)->get('/task/finished');
+        $response = $this->actingAs($this->user)->get('/profile/1');
         $response->assertStatus(200);
-        $response->assertViewHasAll(['tasks', 'priorities']);
+        $response->assertViewHasAll(['userProgressAreaspline', 'user']);
     }
 }
