+++ tests/Feature/Routes/TaskRoutesTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Tests\Unit\Routes;
+namespace Tests\Feature\Routes;
 
-use App\User;
 use Tests\TestCase;
+use App\User;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
-class HomeRoutesTest extends TestCase
+class TaskRoutesTest extends TestCase
 {
 
     use DatabaseTransactions;
@@ -35,11 +35,11 @@ class HomeRoutesTest extends TestCase
      *
      * @test
      */
-    public function it_can_access_the_home_page()
+    public function it_can_access_the_task_page()
     {
-        $response = $this->actingAs($this->user)->get('/home');
+        $response = $this->actingAs($this->user)->get('/task');
         $response->assertStatus(200);
-        $response->assertViewHasAll(['myProjects', 'userProjects', 'userProjectCount', 'userProgressAreaspline']);
+        $response->assertViewHasAll(['tasks', 'priorities']);
     }
 
     /**
@@ -47,11 +47,11 @@ class HomeRoutesTest extends TestCase
      *
      * @test
      */
-    public function it_can_access_the_home_project_page()
+    public function it_can_access_the_unfinished_task_page()
     {
-        $response = $this->actingAs($this->user)->get('/home/project');
+        $response = $this->actingAs($this->user)->get('/task/unfinished');
         $response->assertStatus(200);
-        $response->assertViewHasAll(['myProjects', 'userProjects', 'userProjectCount', 'userProgressAreaspline']);
+        $response->assertViewHasAll(['tasks', 'priorities']);
     }
 
     /**
@@ -59,10 +59,10 @@ class HomeRoutesTest extends TestCase
      *
      * @test
      */
-    public function it_can_access_the_user_profile_page()
+    public function it_can_access_the_finished_task_page()
     {
-        $response = $this->actingAs($this->user)->get('/profile/1');
+        $response = $this->actingAs($this->user)->get('/task/finished');
         $response->assertStatus(200);
-        $response->assertViewHasAll(['userProgressAreaspline', 'user']);
+        $response->assertViewHasAll(['tasks', 'priorities']);
     }
 }
